Use a better system for getting stops.js
The new script, get_stops.py, calls the API directly so I don't have to
remember which API method I need to call, what arguments to use to curl,
etc. Run the script with the API key and it just prints the list of
stops. The list of stops is now formatted so that each stop takes one
line; this is a much more diff'able format and saves 4 kB (!) in
whitespace.